MR-262: fixed issues in routing
