+++ src/Components/Layout/index.jsx
@@ -1,34 +1,15 @@
-import { useSelector } from 'react-redux';
-import { ClipLoader } from 'react-spinners';
-import AdminView from 'Components/Views/adminView';
-import MemberView from 'Components/Views/memberView';
-import SuperAdminView from 'Components/Views/superAdminView';
-import TrainerView from 'Components/Views/trainerView';
-import Container from 'Components/Shared/Container';
-import LandingView from 'Components/Views/landingView';
+import Header from 'Components/Shared/Header';
+import Sidebar from 'Components/Shared/Sidebar';
+import styles from './layout.module.css';
 
-const Layout = () => {
-  const auth = useSelector((state) => state.auth);
-  const role = sessionStorage.getItem('role');
-
-  if (auth?.isAuthPending || auth?.isLoading) {
-    <Container>
-      <ClipLoader />
-    </Container>;
-  }
-
-  switch (role) {
-    case 'SUPER_ADMIN':
-      return <SuperAdminView />;
-    case 'ADMIN':
-      return <AdminView />;
-    case 'TRAINER':
-      return <TrainerView />;
-    case 'MEMBER':
-      return <MemberView />;
-    default:
-      return <LandingView />;
-  }
+const Layout = (props) => {
+  return (
+    <>
+      <Header profileRoute={props.profileRoute} />
+      <Sidebar routes={props.routes} rol={props.rol} />
+      <div className={styles.center}>{props.children}</div>
+    </>
+  );
 };
 
 export default Layout;
