+++ src/Components/Login/login.module.css
@@ -13,12 +13,13 @@
   flex-direction: column;
   align-content: center;
   align-self: center;
+  justify-content: center;
   background-color: #fff;
   border: none;
   box-shadow: 0px 11px 15px -7px rgba(0,0,0,0.2), 0px 24px 38px 3px rgba(0,0,0,0.14), 0px 9px 46px 8px rgba(0,0,0,0.12);
   width: 35vw;
   border-radius: 10px;
-  height: 80vh;
+  height: 70vh;
 }
 
 .big  h2 {
@@ -29,7 +30,7 @@
 .logo {
   width: 15vw;
   cursor: pointer;
-  margin: 5vh auto 0 auto;
+  margin: 0 auto;
 }
 
 .formLogin {
