+++ src/Components/Shared/Container/index.jsx
@@ -1,8 +1,18 @@
 import React from 'react';
 import styles from './container.module.css';
 
-const Container = ({ children, isLogin, center }) => (
-  <div className={isLogin ? styles.login : center ? styles.center : styles.container}>
+const Container = ({ children, isLogin, center, realCenter }) => (
+  <div
+    className={
+      isLogin
+        ? styles.login
+        : center
+        ? styles.center
+        : realCenter
+        ? styles.realCenter
+        : styles.container
+    }
+  >
     {children}
   </div>
 );
