+++ src/Components/SignUp/index.jsx
@@ -68,12 +68,12 @@ const SignUp = () => {
             alt=""
           ></img>
         </div>
-        <div className={styles.containerForm}>
-          <div className={styles.head}>
-            <div id="sign-up-form-go-back" className={styles.arrow} onClick={handleCancel}>
-              <FiArrowLeft size={25} />
-            </div>
+        <div className={styles.head}>
+          <div id="sign-up-form-go-back" className={styles.arrow} onClick={handleCancel}>
+            <FiArrowLeft size={25} />
           </div>
+          <div className={styles.containerForm}></div>
+
           <form onSubmit={handleSubmit(onSubmit)}>
             <div className={styles.formMembers}>
               <div className={styles.formColumn}>
