+++ src/Components/SignUp/signup.module.css
@@ -21,19 +21,18 @@
 }
 
 .containerForm {
-  flex: 1;
   display: flex;
   flex-direction: column;
-  align-items: center;
   max-width: 50vw;
+  justify-content: center;
   padding: 2vh 2vw;
 }
 
 .head {
   display: flex;
   flex-direction: column;
-  align-items: flex-start;
-  width: 100% ;
+  width: 50vw;
+  margin-top: 30px;
 }
 
 .formMembers {
@@ -54,7 +53,7 @@
 }
 
 .arrow {
-  margin-left: 1vw;
+  margin-left: 31.76px;
   cursor: pointer;
 }
