--- src/Components/Views/landingView.jsx
@@ -1,14 +1,0 @@
-import styles from 'Components/LandingPage/landing.module.css';
-import Routes from 'Routes/Routes';
-
-const LandingView = () => {
-  return (
-    <>
-      <div className={styles.body}>
-        <Routes />
-      </div>
-    </>
-  );
-};
-
-export default LandingView;
