--- src/Components/Views/loginView.jsx
@@ -1,14 +1,0 @@
-import styles from 'Components/Layout/layout.module.css';
-import Routes from 'Routes/Routes';
-
-const LoginView = () => {
-  return (
-    <>
-      <div className={styles.body}>
-        <Routes />
-      </div>
-    </>
-  );
-};
-
-export default LoginView;
