+++ src/index.js
@@ -1,17 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Layout from 'Components/Layout';
 import { BrowserRouter } from 'react-router-dom';
 import './index.css';
-
 import { Provider } from 'react-redux';
 import store from 'Redux/store';
+import Routes from 'Routes/Routes';
 
 ReactDOM.render(
   <Provider store={store}>
     <React.StrictMode>
       <BrowserRouter>
-        <Layout />
+        <Routes />
       </BrowserRouter>
     </React.StrictMode>
   </Provider>,
